Changes to reduce warnings & linker errors in Windows builds 1. casts (typically to int) in various member functions    of Alignment. 2. Use of ctime_s instead of ctime in reportPhyloAnalysis()     (conditional, only takes effect in MSVC builds) 3. BlockAllocator::allocateVectorOfParsimonyBlocks, first     parameter is now an intptr_t rather than an int. 4. Use of sprintf_s rather than sprintf in Node::Node     (conditional, only takes effect in MSVC builds) 5. Changed types (parsimony_score now double),     (iteration now intptr_t) in PhyloTree::doParsimonySearch() 6. Changed type of ParsimonySearchParameters::radius.     It is now int (as if it would make sense to ask for a radius     in the billions!). 7. Added some static casts in PhyloTree::mergeAlignment(). 8. Likewise, PhyloTree::setNumThreads() (which now,     won't set more than the number allowed by OPEN MP) 9. in tree/quartet.cpp, initeps() calls localtime_s and     asc_time_s (if compiled under MSVC). 10. Likewise, in  sprng/makeseed.c, localtime_s is used in       place of localtime() (if compiled under MSVC). 11. sprng/lcg64.c only #defines NDEBUG if it is not already       defined (formerly it was unconditional). 12. Changed the way that /MD is replaced with /MT,        and /MDd is replaced with /MTd, in pure-MSVC builds.        (In the main CMakeLists.txt). 13. MTree::scaleCladeSupport doesn't rely on an       non-standard std::string:operator+=(double).       It uses a std::stringstream to format the number.
Following changes are for MPI in Windows builds
14. Dynamically-sized arays in functions in the pll file
     genericParallelization.c, are now acquired via malloc
     (in MSVC builds).  In:
     (a) broadCastAlpha
     (b) broadCastLg4xWeights
     (c) copyLG4
     (d) broadCastRates
     (e) execFunction
     (f) distributeYVectors

+++ placement/blockallocator.cpp
@@ -25,9 +25,9 @@ void BlockAllocator::allocateParsimonyBlock(UINT*& partial_pars) {
     ASSERT( partial_pars < phylo_tree.tip_partial_pars );
     ++index_parsimony;
 }
-void BlockAllocator::allocateVectorOfParsimonyBlocks(int vectorSize,
+void BlockAllocator::allocateVectorOfParsimonyBlocks(intptr_t vectorSize,
     std::vector<UINT*>& buffers) {
-    for (int i = 0; i < vectorSize; ++i) {
+    for (intptr_t i = 0; i < vectorSize; ++i) {
         size_t offset = index_parsimony * phylo_tree.pars_block_size;
         buffers.push_back(phylo_tree.central_partial_pars + offset);
         ASSERT(buffers.back() < phylo_tree.tip_partial_pars);
